requirements.txt
appdirs==1.4.4
asgiref==3.3.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
distlib==0.3.1
Django==3.1.6
django-taggit==1.4.0
django-tinymce==3.3.0
filelock==3.0.12
Flask==1.1.2
gunicorn==20.1.0
html5lib==1.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
nose==1.1.2
Pillow==8.1.0
psycopg2-binary==2.8.6
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.4
virtualenv==20.4.3
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.2.0
